fix: avoid auth errors during javadoc publish step

+++ build/publish-javadoc.sh
@@ -1,23 +1,20 @@
 #!/bin/bash
 
-# We'll publish javadocs ONLY for a tagged release build.
-if [ -n "$TRAVIS_TAG" ]; then
+# Avoid publishing javadocs for a PR build
+if [ "$TRAVIS_PULL_REQUEST" == "false" ] && [ "$TRAVIS_BRANCH" ]; then
 
     printf "\n>>>>> Publishing javadoc for release build: repo=%s branch=%s build_num=%s job_num=%s\n" ${TRAVIS_REPO_SLUG} ${TRAVIS_BRANCH} ${TRAVIS_BUILD_NUMBER} ${TRAVIS_JOB_NUMBER} 
 
-    git config --global user.email "devxsdk@us.ibm.com"
-    git config --global user.name "IBM Cloud SDK Developement"
-
     printf "\n>>>>> Cloning repository's gh-pages branch into directory 'gh-pages'"
-    git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/IBM/platform-services-java-sdk.git gh-pages
+    git clone --branch=gh-pages https://${GH_TOKEN}@github.com/IBM/platform-services-java-sdk.git gh-pages
 
-    printf "\n>>>>> Finished cloning..."
+    printf "\n>>>>> Finished cloning...\n"
 
     
     pushd gh-pages
     
-    # Create a new directory for this tagged release and copy the aggregated javadocs there.
-    printf "\n>>>>> Copying aggregated javadocs to new tagged-release directory: %s\n" ${TRAVIS_TAG}
+    # Create a new directory for this branch/tag and copy the aggregated javadocs there.
+    printf "\n>>>>> Copying aggregated javadocs to new tagged-release directory: %s\n" ${TRAVIS_BRANCH}
     rm -rf docs/${TRAVIS_BRANCH}
     mkdir -p docs/${TRAVIS_BRANCH}
     cp -rf ../target/site/apidocs/* docs/${TRAVIS_BRANCH}
@@ -25,18 +22,20 @@ if [ -n "$TRAVIS_TAG" ]; then
     printf "\n>>>>> Generating gh-pages index.html...\n"
     ../build/generate-index-html.sh > index.html
 
-    # Update the latest symlink to point to this new tag
-    pushd docs
-    rm latest
-    ln -s ./${TRAVIS_TAG} latest
-    printf "\n>>>>> Updated 'docs/latest' symlink:\n"
-    ls -l latest
-    popd
+    # Update the 'latest' symlink to point to this branch if it's a tagged release.
+    if [ -n "$TRAVIS_TAG" ]; then
+	pushd docs
+	rm latest
+	ln -s ./${TRAVIS_TAG} latest
+	printf "\n>>>>> Updated 'docs/latest' symlink:\n"
+	ls -l latest
+	popd
+    fi
 
     printf "\n>>>>> Committing new javadoc...\n"
     git add -f .
     git commit -m "Javadoc for release ${TRAVIS_TAG} (${TRAVIS_COMMIT})"
-    git push -fq origin gh-pages
+    git push -f origin gh-pages
 
     popd
